+++ apps/journeys-admin/src/components/LanguageSwitcher/index.ts
@@ -1,0 +1,1 @@
+export { LanguageSwitcher } from './LanguageSwitcher'
